+++ web/src/category-metadata.js
@@ -3,74 +3,85 @@ const PAGES = require('../../ui/constants/pages');
 // Uncomment as you add metadata
 
 module.exports.CATEGORY_METADATA = {
-  [PAGES.BIG_HITS]: {
+  [PAGES.BIG_HITS]: () => ({
     title: 'Big Hits',
     description: 'Animation, pop culture, comedy, and all the other weird on Odysee',
     image: '',
-  },
-  // [PAGES.CREATIVE_ARTS]: {
-  //   title: 'Creative Arts',
-  //   description: '',   <----- FILL IN PLEASE
-  //   image: '',         <----- FILL IN PLEASE
-  // },
-  [PAGES.COMMUNITY]: {
+  }),
+  [PAGES.COMMUNITY]: () => ({
     title: 'The Universe',
     description: 'Podcasts, life, learning, and everything else on Odysee',
     image: '',
-  },
-  //   [PAGES.ENLIGHTENMENT]: {
+  }),
+  // [PAGES.CREATIVE_ARTS]: () => ({
+  //   title: 'Creative Arts',
+  //   description: '',   <----- FILL IN PLEASE
+  //   image: '',         <----- FILL IN PLEASE
+  // }),
+  //   [PAGES.ENLIGHTENMENT]: () => ({
   //     title: '',
   //     description: '',
   //     image: '',
-  //   },
-  [PAGES.FINANCE]: {
+  //   }),
+  [PAGES.FINANCE]: () => ({
     title: 'Finance 2.0',
     description: 'Crypto, Money, Economics, Markets on Odysee ',
     image: 'https://cdn.lbryplayer.xyz/speech/category-finance:c.jpg',
-  },
-  [PAGES.GAMING]: {
+  }),
+  [PAGES.GAMING]: () => ({
     title: 'Gaming',
     description: 'Pew pew bzzz gaming on Odysee',
     image: 'https://cdn.lbryplayer.xyz/speech/category-gaming:5.jpg',
-  },
-  [PAGES.GENERAL]: {
+  }),
+  [PAGES.GENERAL]: () => ({
     title: 'Cheese',
     description: 'Cheese is the answer to life, the universe, and everything. We have primo cheese on Odysee',
     image: 'https://cdn.lbryplayer.xyz/speech/category-primary1:5.jpg',
-  },
-  [PAGES.LAB]: {
+  }),
+  [PAGES.LAB]: () => ({
     title: 'Lab',
     description: 'Science - the real kind, on Odysee',
     image: '',
-  },
-  [PAGES.NEWS]: {
+  }),
+  [PAGES.NEWS]: () => ({
     title: 'News & Politics',
     description: `Stay up to date with all that's happening around the world on Odysee`,
     image: '',
-  },
-  [PAGES.MOVIES]: {
+  }),
+  [PAGES.MOVIES]: () => ({
     title: 'Movies',
     description: `Do you love B rated movies? We've got you covered on Odysee`,
     image: 'https://cdn.lbryplayer.xyz/speech/category-movies:2.jpg',
-  },
-  [PAGES.MUSIC]: {
+  }),
+  [PAGES.MUSIC]: () => ({
     title: 'Music',
     description: 'All the songs, reviews, covers, and how-tos you love on Odysee',
     image: 'https://cdn.lbryplayer.xyz/speech/category-music:8.jpg',
+  }),
+  [PAGES.SEARCH]: ({ q = '' }) => {
+    if (!q) {
+      return {};
+    }
+
+    return {
+      title: `"${q}" Search Results`,
+      description: `Find the best "${q}" content on Odysee`,
+      urlQueryString: `q=${q}`,
+    };
   },
-  [PAGES.TECH]: {
+  [PAGES.TECH]: () => ({
     title: 'Tech',
     description: 'Hardware, software, startups, photography on Odysee',
     image: '',
-  },
-  [PAGES.TECHNOLOGY]: {
+  }),
+  [PAGES.TECHNOLOGY]: () => ({
     title: 'Tech',
     description: 'Hardware, software, startups, photography on Odysee',
     image: '',
-  },
-  [PAGES.WILD_WEST]: {
+  }),
+  [PAGES.WILD_WEST]: () => ({
     title: 'Wild West',
     description: 'Boosted by user credits, this is what the community promotes on Odysee',
     image: 'https://cdn.lbryplayer.xyz/speech/category-wildwest:1.jpg',
-  },
+  }),
 };