+++ web/src/html.js
@@ -55,7 +55,7 @@ function truncateDescription(description, maxChars = 200) {
   return chars.length > maxChars ? truncated + '...' : truncated;
 }
 
-function getCategoryMeta(path) {
+function getCategoryMetaRenderFn(path) {
   const page = Object.keys(CATEGORY_METADATA).find((x) => path === `/$/${x}` || path === `/$/${x}/`);
   return CATEGORY_METADATA[page];
 }
@@ -64,14 +64,16 @@ function getCategoryMeta(path) {
 // Normal metadata with option to override certain values
 //
 function buildOgMetadata(overrideOptions = {}) {
-  const { title, description, image, path } = overrideOptions;
-  const cleanDescription = removeMd(description || SITE_DESCRIPTION);
+  const { title, description, image, path, urlQueryString } = overrideOptions;
+  const cleanDescription = escapeHtmlProperty(removeMd(description || SITE_DESCRIPTION));
+  const cleanTitle = escapeHtmlProperty(title);
+  const url = (path ? `${URL}${path}` : URL) + (urlQueryString ? `?${urlQueryString}` : '');
 
   const head =
     `<title>${SITE_TITLE}</title>\n` +
     `<meta name="description" content="${cleanDescription}" />\n` +
-    `<meta property="og:url" content="${path ? `${URL}${path}` : URL}" />\n` +
-    `<meta property="og:title" content="${title || OG_HOMEPAGE_TITLE || SITE_TITLE}" />\n` +
+    `<meta property="og:url" content="${url}" />\n` +
+    `<meta property="og:title" content="${cleanTitle || OG_HOMEPAGE_TITLE || SITE_TITLE}" />\n` +
     `<meta property="og:site_name" content="${SITE_NAME || SITE_TITLE}"/>\n` +
     `<meta property="og:description" content="${cleanDescription}" />\n` +
     `<meta property="og:image" content="${
@@ -80,7 +82,7 @@ function buildOgMetadata(overrideOptions = {}) {
     `<meta property="og:type" content="website"/>\n` +
     '<meta name="twitter:card" content="summary_large_image"/>\n' +
     `<meta name="twitter:title" content="${
-      (title && title + ' ' + OG_TITLE_SUFFIX) || OG_HOMEPAGE_TITLE || SITE_TITLE
+      (cleanTitle && cleanTitle + ' ' + OG_TITLE_SUFFIX) || OG_HOMEPAGE_TITLE || SITE_TITLE
     }" />\n` +
     `<meta name="twitter:description" content="${cleanDescription}" />\n` +
     `<meta name="twitter:image" content="${
@@ -345,12 +347,11 @@ async function getHtml(ctx) {
     return insertToHead(html);
   }
 
-  const categoryMeta = getCategoryMeta(requestPath);
-  if (categoryMeta) {
+  const categoryMetaFn = getCategoryMetaRenderFn(requestPath);
+  if (categoryMetaFn) {
+    const categoryMeta = categoryMetaFn(ctx.request.query);
     const categoryPageMetadata = buildOgMetadata({
-      title: categoryMeta.title,
-      description: categoryMeta.description,
-      image: categoryMeta.image,
+      ...categoryMeta,
       path: requestPath,
     });
     return insertToHead(html, categoryPageMetadata);